+++ docs/reference/changelog.rst
@@ -28,7 +28,8 @@ New features:
 
 - Add a ``gradle-use`` plugin for publishing Gradle artifacts to a local Maven
   repository.
-- Allow ``build-snaps`` to use the ``=`` separator between the snap name and channel.
+- Allow ``stage-snaps`` and ``build-snaps`` to use the ``@`` separator between the
+  snap name and channel.
 
 .. _release-2.33.0: